+++ config/development.js
@@ -2,5 +2,6 @@ export const Development = {
   API_HOST: 'https://staging.api.pokedextracker.com',
   ENVIRONMENT: 'development',
   GA_ID: 'UA-45307701-5',
-  SENTRY_DSN: 'https://f2362849081c4b2494b6084fec02300e@sentry.io/108494'
+  SENTRY_DSN: 'https://f2362849081c4b2494b6084fec02300e@sentry.io/108494',
+  STRIPE_PUBLISHABLE_KEY: 'pk_test_OSgFpGR8zFRl4d1PRS8e3icU'
 };
