+++ config/production.js
@@ -2,5 +2,6 @@ export const Production = {
   API_HOST: 'https://api.pokedextracker.com',
   ENVIRONMENT: 'production',
   GA_ID: 'UA-45307701-4',
-  SENTRY_DSN: 'https://dc4726f104294d3ebe692a9df024dcb5@sentry.io/103004'
+  SENTRY_DSN: 'https://dc4726f104294d3ebe692a9df024dcb5@sentry.io/103004',
+  STRIPE_PUBLISHABLE_KEY: 'pk_live_vSUhWqIZtwt7s4NguleOGtJ5'
 };
