+++ package.json
@@ -32,6 +32,7 @@
     "react-redux": "^5.0.6",
     "react-router": "^3.0.2",
     "react-router-redux": "^4.0.8",
+    "react-stripe-elements": "^1.2.0",
     "redux": "^3.6.0",
     "redux-thunk": "^2.2.0",
     "slug": "^0.9.1"
