+++ public/index.html
@@ -43,6 +43,7 @@
         document.documentElement.className += ' no-touch';
       }
     </script>
+    <script src="https://js.stripe.com/v3/"></script>
   </head>
   <body>
     <div id="root"></div>
